libc: refactor socket operations
